Added Cube class to plot a Cube
I couldn't figure out how to plot multiple cubes so far so this will just plot a random cube rn. I've done a bunch of tutorials and have code on my computer that plots multiple cubes but I haven't been able to figure out how to make it work for this yet.